Dynamically loading the spec at the last second through an ajax request from a folder that can be set as a volume when starting the container like so:
docker run \
-d \
-p 80:8080 \
--volume $HOME/spec:/build/dist/spec \
swagger-ui-builder

This means that one can then keep changing the contents of a swagger-spec.json file inside the host volume folder (in this case at $HOME/spec/swagger-json) and it will immediately take effect without having to restart the container. If the user does not provide a swagger-spec.json file then the default petstore is shown as normal.